Revert "Merge branch 'backend_refatoracoes2' into Front"
This reverts commit 04294f19b711f31465e583b30a1bfc6da4f47643.

+++ aba/src/main/java/com/aba/service/AlunoServiceImpl.java
@@ -3,11 +3,12 @@ import java.util.List;
 import java.util.Optional;
 
 import com.aba.dto.AlunosDTO;
-import com.aba.excecoes.AlunoInexistenteException;
 import com.aba.interfaces.InstrutorService;
 import com.aba.interfaces.PlanoObjetivosService;
+import com.aba.model.Instrutor;
 import com.aba.model.PlanoObjetivos;
-import com.aba.util.erros.ErroUsuario;
+import com.aba.util.CustomErrorType;
+import com.aba.util.MessageError;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -36,12 +37,14 @@ public class AlunoServiceImpl implements AlunoService {
     PlanoObjetivosService planoObjetivosService;
 
     public ResponseEntity<?> cadastrarAluno(AlunoDTO alunoDTO) {
-       Usuario aluno;
+        Instrutor instrutor = this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor());
+        if(instrutor == null){
+            return MessageError.erroInstrutorNaoEncontradoByEmail(alunoDTO.getEmailInstrutor());
+        }
 
-       // o aluno pode tentar encontrar o instrutor e não conseguir
-        aluno = new Aluno(alunoDTO.getNome(), alunoDTO.getDataNascimento(), alunoDTO.getCid(),
-                this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor()),
-                alunoDTO.getContato(), alunoDTO.getGenero(),
+        Usuario aluno;
+        aluno = new Aluno(alunoDTO.getNome(), alunoDTO.getIdade(),
+                instrutor, alunoDTO.getContato(), alunoDTO.getGenero(),
                 alunoDTO.getCpf(), alunoDTO.getEndereco(),
                 alunoDTO.getResponsavel(), alunoDTO.getParentesco());
 
@@ -51,13 +54,12 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.CREATED).body(((Aluno) aluno).getDtoCompleto());
     }
 
-    public ResponseEntity<?> editarAluno(Long id, AlunoDTO alunoDTO) {
-        Aluno aluno;
-        try {
-            aluno = this.getAlunoById(id);
-        } catch (AlunoInexistenteException e) {
-            return ErroUsuario.alunoInexistente(id);
+    public ResponseEntity<?> editarAluno(Long id, AlunoDTO alunoDTO){
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
         }
+
         aluno.editar(alunoDTO, this.instrutorService.getInstrutorByEmail(alunoDTO.getEmailInstrutor()));
         this.alunoRepository.save(aluno);
 
@@ -65,6 +67,10 @@ public class AlunoServiceImpl implements AlunoService {
     }
 
     public ResponseEntity<?> removerAluno(Long id) {
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
+        }
         this.usuarioRepository.deleteById(id);
 
         return ResponseEntity.status(HttpStatus.OK).body("Aluno(a) removido(a)!");
@@ -78,16 +84,15 @@ public class AlunoServiceImpl implements AlunoService {
     }
 
     public ResponseEntity<?> consultarAluno(Long id) {
-        Aluno aluno;
-        try {
-            aluno = this.getAlunoById(id);
-        } catch (AlunoInexistenteException e) {
-            return ErroUsuario.alunoInexistente(id);
+        Aluno aluno = this.getAlunoById(id);
+        if(aluno == null){
+            return MessageError.erroAlunoNaoEncontrado(id);
         }
+
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDto());
     }
 
-    public ResponseEntity<?> adicionarPlanoObjetivos(Long id, Long idPlano) throws AlunoInexistenteException {
+    public ResponseEntity<?> adicionarPlanoObjetivos(Long id, Long idPlano){
         Aluno aluno;
         aluno = getAlunoById(id);
         PlanoObjetivos planoObjetivos;
@@ -99,7 +104,7 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDtoCompleto());
     }
 
-    public ResponseEntity<?> removerPlanoObjetivos(Long id, Long idPlano) throws AlunoInexistenteException {
+    public ResponseEntity<?> removerPlanoObjetivos(Long id, Long idPlano){
         Aluno aluno;
         aluno = getAlunoById(id);
         PlanoObjetivos planoObjetivos;
@@ -111,13 +116,12 @@ public class AlunoServiceImpl implements AlunoService {
         return ResponseEntity.status(HttpStatus.OK).body(aluno.getDtoCompleto());
     }
 
-    public Aluno getAlunoById(Long id) throws AlunoInexistenteException {
+    public Aluno getAlunoById(Long id){
         Optional<Aluno> alunoOptional = this.alunoRepository.findById(id);
 
         if (!alunoOptional.isPresent()) {
-            throw new AlunoInexistenteException("O aluno com id mencionado não existe");
-        }
-
-        return alunoOptional.get();
+            return null;
+        }return alunoOptional.get();
     }
+
 }
